topics/web/serializers: Expand JSON examples
- Give a basic example of JSON before we introduce the one with all of
the struct tags
- Show both encoding and decoding
- Also show the []byte marshaling and unmarshaling. We need it some
time. I intend to describe the difference as roughly "when are working
with a []byte use Marshal. When you're working with io.Readers and
io.Writers use decoders"
- Give an example of how to connect json to net/http. Previously this
just showed up and left people confused how they went together
