+++ topics/web/serializers/README.md
@@ -6,15 +6,17 @@ Learn the basics of using and applying custom JSON and XML serializers.
 
 https://golang.org/pkg/encoding/json  
 https://golang.org/pkg/encoding/xml  
-mholt.github.io/json-to-go/  
+http://mholt.github.io/json-to-go/  
 
 ## Code Review
 
-JSON Encoding: [Code](example1/main.go) | [Test](example1/main_test.go)  
+JSON Encoding: [Code](example1/main.go)  
+JSON Struct Tags: [Code](example2/main.go) | [Test](example2/main_test.go)  
 MarshalJSON Interface: [Code](example2/main.go) | [Test](example2/main_test.go)  
-XML Encoding: [Code](example3/main.go) | [Test](example3/main_test.go)  
-MarshalXML Interface: [Code](example4/main.go) | [Test](example4/main_test.go)  
-JSON Mistakes: [Code](example5/main.go)  
+JSON Mistakes: [Code](example4/main.go)  
+JSON and net/http: [Code](example5/main.go)  
+XML Encoding: [Code](example6/main.go) | [Test](example6/main_test.go)  
+MarshalXML Interface: [Code](example7/main.go) | [Test](example7/main_test.go)  
 
 ## Exercises
