+++ topics/web/serializers/example2/main_test.go
@@ -1,8 +1,8 @@
 // All material is licensed under the Apache License Version 2.0, January 2004
 // http://www.apache.org/licenses/LICENSE-2.0
 
-// Tests for the sample program to show how to implement the
-// json.Marshaler interface to dictate the marshaling.
+// Tests for the sample program to show how to control JSON encoding using
+// struct tags.
 package main
 
 import (
@@ -26,7 +26,7 @@ func TestEncodeZeroValueUser(t *testing.T) {
 
 	// Validate we received the expected response.
 	got := strings.TrimSpace(bb.String())
-	want := `{"Admin":false,"Bio":null,"CreatedAt":"0001-01-01T00:00:00Z","first_name":""}`
+	want := `{"first_name":"","CreatedAt":"0001-01-01T00:00:00Z","Admin":false,"Bio":null}`
 	if got != want {
 		t.Log("Wanted:", want)
 		t.Log("Got   :", got)
@@ -53,7 +53,7 @@ func TestEncodeUser(t *testing.T) {
 
 	// Validate we received the expected response.
 	got := strings.TrimSpace(bb.String())
-	want := `{"Admin":false,"Bio":null,"CreatedAt":"0001-01-01T00:00:00Z","LastName":"Jane","first_name":"Mary"}`
+	want := `{"first_name":"Mary","LastName":"Jane","CreatedAt":"0001-01-01T00:00:00Z","Admin":false,"Bio":null}`
 	if got != want {
 		t.Log("Wanted:", want)
 		t.Log("Got   :", got)
