+++ topics/web/serializers/example3/main_test.go
@@ -1,12 +1,13 @@
 // All material is licensed under the Apache License Version 2.0, January 2004
 // http://www.apache.org/licenses/LICENSE-2.0
 
-// Tests for the sample program to show how to use the XML encoder.
+// Tests for the sample program to show how to implement the
+// json.Marshaler interface to dictate the marshaling.
 package main
 
 import (
 	"bytes"
-	"encoding/xml"
+	"encoding/json"
 	"strings"
 	"testing"
 )
@@ -16,16 +17,16 @@ func TestEncodeZeroValueUser(t *testing.T) {
 	// Create a bytes buffer for our writer.
 	var bb bytes.Buffer
 
-	// Encode a zero value user and write the XML
+	// Encode a zero value user and write the JSON
 	// to the bytes buffer.
-	err := xml.NewEncoder(&bb).Encode(&User{})
+	err := json.NewEncoder(&bb).Encode(&User{})
 	if err != nil {
 		t.Fatal(err)
 	}
 
 	// Validate we received the expected response.
 	got := strings.TrimSpace(bb.String())
-	want := `<User><first_name></first_name><CreatedAt>0001-01-01T00:00:00Z</CreatedAt><Admin>false</Admin></User>`
+	want := `{"Admin":false,"Bio":null,"CreatedAt":"0001-01-01T00:00:00Z","first_name":""}`
 	if got != want {
 		t.Log("Wanted:", want)
 		t.Log("Got   :", got)
@@ -44,15 +45,15 @@ func TestEncodeUser(t *testing.T) {
 		LastName:  "Jane",
 	}
 
-	// Encode the user and write the XML to the bytes buffer.
-	err := xml.NewEncoder(&bb).Encode(&u)
+	// Encode the user and write the JSON to the bytes buffer.
+	err := json.NewEncoder(&bb).Encode(&u)
 	if err != nil {
 		t.Fatal(err)
 	}
 
 	// Validate we received the expected response.
 	got := strings.TrimSpace(bb.String())
-	want := `<User><first_name>Mary</first_name><LastName>Jane</LastName><CreatedAt>0001-01-01T00:00:00Z</CreatedAt><Admin>false</Admin></User>`
+	want := `{"Admin":false,"Bio":null,"CreatedAt":"0001-01-01T00:00:00Z","LastName":"Jane","first_name":"Mary"}`
 	if got != want {
 		t.Log("Wanted:", want)
 		t.Log("Got   :", got)
