+++ topics/web/serializers/example6/main.go
@@ -1,11 +1,12 @@
 // All material is licensed under the Apache License Version 2.0, January 2004
 // http://www.apache.org/licenses/LICENSE-2.0
 
-// Sample program to show how to control JSON encoding using struct tags.
+// Sample program to show how to use the XML encoder.
 package main
 
 import (
-	"encoding/json"
+	"encoding/xml"
+	"fmt"
 	"log"
 	"os"
 	"time"
@@ -14,9 +15,9 @@ import (
 // User represents a user in the system. Omit the LastName
 // from Marshaling unless we have a value. Omit Age always.
 type User struct {
-	FirstName string `json:"first_name"`
-	LastName  string `json:",omitempty"`
-	Age       int    `json:"-"`
+	FirstName string `xml:"first_name"`
+	LastName  string `xml:",omitempty"`
+	Age       int    `xml:"-"`
 	CreatedAt time.Time
 	Admin     bool
 	Bio       *string
@@ -25,10 +26,12 @@ type User struct {
 func main() {
 
 	// Encode a zero valued version of a user and write to stdout.
-	err := json.NewEncoder(os.Stdout).Encode(&User{})
+	err := xml.NewEncoder(os.Stdout).Encode(&User{})
 	if err != nil {
 		log.Fatal(err)
 	}
+
+	fmt.Println()
 
 	// Create a user value for Mary Jane.
 	u := User{
@@ -37,7 +40,7 @@ func main() {
 	}
 
 	// Encode the user value and write to stdout.
-	err = json.NewEncoder(os.Stdout).Encode(&u)
+	err = xml.NewEncoder(os.Stdout).Encode(&u)
 	if err != nil {
 		log.Fatal(err)
 	}
