+++ topics/web/serializers/example6/main_test.go
@@ -1,12 +1,12 @@
 // All material is licensed under the Apache License Version 2.0, January 2004
 // http://www.apache.org/licenses/LICENSE-2.0
 
-// Tests for the sample program to show how to use the JSON encoder.
+// Tests for the sample program to show how to use the XML encoder.
 package main
 
 import (
 	"bytes"
-	"encoding/json"
+	"encoding/xml"
 	"strings"
 	"testing"
 )
@@ -16,16 +16,16 @@ func TestEncodeZeroValueUser(t *testing.T) {
 	// Create a bytes buffer for our writer.
 	var bb bytes.Buffer
 
-	// Encode a zero value user and write the JSON
+	// Encode a zero value user and write the XML
 	// to the bytes buffer.
-	err := json.NewEncoder(&bb).Encode(&User{})
+	err := xml.NewEncoder(&bb).Encode(&User{})
 	if err != nil {
 		t.Fatal(err)
 	}
 
 	// Validate we received the expected response.
 	got := strings.TrimSpace(bb.String())
-	want := `{"first_name":"","CreatedAt":"0001-01-01T00:00:00Z","Admin":false,"Bio":null}`
+	want := `<User><first_name></first_name><CreatedAt>0001-01-01T00:00:00Z</CreatedAt><Admin>false</Admin></User>`
 	if got != want {
 		t.Log("Wanted:", want)
 		t.Log("Got   :", got)
@@ -44,15 +44,15 @@ func TestEncodeUser(t *testing.T) {
 		LastName:  "Jane",
 	}
 
-	// Encode the user and write the JSON to the bytes buffer.
-	err := json.NewEncoder(&bb).Encode(&u)
+	// Encode the user and write the XML to the bytes buffer.
+	err := xml.NewEncoder(&bb).Encode(&u)
 	if err != nil {
 		t.Fatal(err)
 	}
 
 	// Validate we received the expected response.
 	got := strings.TrimSpace(bb.String())
-	want := `{"first_name":"Mary","LastName":"Jane","CreatedAt":"0001-01-01T00:00:00Z","Admin":false,"Bio":null}`
+	want := `<User><first_name>Mary</first_name><LastName>Jane</LastName><CreatedAt>0001-01-01T00:00:00Z</CreatedAt><Admin>false</Admin></User>`
 	if got != want {
 		t.Log("Wanted:", want)
 		t.Log("Got   :", got)
